Refactoring: split Dashboard & DashboardSiaes
