+++ lemarche/www/dashboard/views.py
@@ -1,40 +1,15 @@
-from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.messages.views import SuccessMessageMixin
-from django.http import HttpResponseRedirect
-from django.shortcuts import get_object_or_404
 from django.urls import reverse_lazy
 from django.utils import timezone
-from django.utils.safestring import mark_safe
-from django.views.generic import DeleteView, DetailView, ListView, UpdateView
-from django.views.generic.edit import FormMixin
+from django.views.generic import DetailView, UpdateView
 
 from lemarche.cms.models import ArticlePage
 from lemarche.cms.snippets import ArticleCategory
-from lemarche.siaes.models import Siae, SiaeUser, SiaeUserRequest
+from lemarche.siaes.models import Siae
 from lemarche.tenders.models import Tender
 from lemarche.users.models import User
-from lemarche.utils.mixins import SiaeMemberRequiredMixin, SiaeUserAndNotMemberRequiredMixin, SiaeUserRequiredMixin
-from lemarche.utils.s3 import S3Upload
-from lemarche.www.dashboard.forms import (
-    ProfileEditForm,
-    SiaeClientReferenceFormSet,
-    SiaeEditContactForm,
-    SiaeEditInfoForm,
-    SiaeEditLinksForm,
-    SiaeEditOfferForm,
-    SiaeEditSearchForm,
-    SiaeImageFormSet,
-    SiaeLabelOldFormSet,
-    SiaeOfferFormSet,
-    SiaeSearchAdoptConfirmForm,
-    SiaeSearchBySiretForm,
-    SiaeUserRequestForm,
-)
-from lemarche.www.dashboard.tasks import (
-    send_siae_user_request_email_to_assignee,
-    send_siae_user_request_response_email_to_initiator,
-)
+from lemarche.www.dashboard.forms import ProfileEditForm
 
 
 SLUG_RESSOURCES_CAT_SIAES = "solutions"
@@ -107,306 +82,3 @@ class ProfileEditView(LoginRequiredMixin, SuccessMessageMixin, UpdateView):
 
     def get_object(self):
         return self.request.user
-
-
-class SiaeSearchBySiretView(SiaeUserRequiredMixin, FormMixin, ListView):
-    form_class = SiaeSearchBySiretForm
-    template_name = "dashboard/siae_search_by_siret.html"
-    context_object_name = "siaes"
-    # queryset =
-
-    def get_queryset(self):
-        """Filter results."""
-        filter_form = SiaeSearchBySiretForm(data=self.request.GET)
-        results = filter_form.filter_queryset()
-        return results
-
-    def get_context_data(self, **kwargs):
-        """
-        - initialize the form with the query parameters (only if they are present)
-        """
-        context = super().get_context_data(**kwargs)
-        if len(self.request.GET.keys()):
-            context["form"] = SiaeSearchBySiretForm(data=self.request.GET)
-        return context
-
-
-class SiaeSearchAdoptConfirmView(SiaeUserAndNotMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeSearchAdoptConfirmForm
-    template_name = "dashboard/siae_search_adopt_confirm.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Rattachement confirmé ! Vous pouvez dès à présent mettre à jour sa fiche."
-
-    def get_context_data(self, **kwargs):
-        """
-        - check if there isn't any pending SiaeUserRequest
-        """
-        context = super().get_context_data(**kwargs)
-        siae_user_pending_request = self.object.siaeuserrequest_set.initiator(self.request.user).pending()
-        context["siae_user_pending_request"] = siae_user_pending_request
-        return context
-
-    def form_valid(self, form):
-        """
-        - Siae with user? add the User to the Siae
-        - Siae with existing user(s)? go through the invitation workflow
-        """
-        if not self.object.users.count():
-            self.object.users.add(self.request.user)
-            return super().form_valid(form)
-        else:
-            # create SiaeUserRequest + send request email to assignee
-            siae_user_request = SiaeUserRequest.objects.create(
-                siae=self.object,
-                initiator=self.request.user,
-                assignee=self.object.users.first(),
-                logs=[{"action": "create", "timestamp": timezone.now().isoformat()}],
-            )
-            send_siae_user_request_email_to_assignee(siae_user_request)
-            success_message = (
-                f"La demande a été envoyée à {self.object.users.first().full_name}.<br />"
-                f"<i>Cet utilisateur ne fait plus partie de la structure ? <a href=\"{reverse_lazy('dashboard:siae_search_by_siret')}?siret={self.object.siret}\">Contactez le support</a></i>"  # noqa
-            )
-            messages.add_message(self.request, messages.SUCCESS, success_message)
-            return HttpResponseRedirect(reverse_lazy("dashboard:home"))
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit", args=[self.kwargs.get("slug")])
-
-
-class SiaeUsersView(SiaeMemberRequiredMixin, DetailView):
-    template_name = "dashboard/siae_users.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-
-    def get_context_data(self, **kwargs):
-        """
-        - check if there isn't any pending SiaeUserRequest
-        """
-        context = super().get_context_data(**kwargs)
-        siae_user_pending_request = self.object.siaeuserrequest_set.assignee(self.request.user).pending()
-        context["siae_user_pending_request"] = siae_user_pending_request
-        return context
-
-
-class SiaeEditSearchView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeEditSearchForm
-    template_name = "dashboard/siae_edit_search.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Vos modifications ont bien été prises en compte."
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit_search", args=[self.kwargs.get("slug")])
-
-
-class SiaeEditInfoView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeEditInfoForm
-    template_name = "dashboard/siae_edit_info.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Vos modifications ont bien été prises en compte."
-
-    def get_context_data(self, **kwargs):
-        """
-        - pass s3 image upload config
-        - init label form
-        """
-        context = super().get_context_data(**kwargs)
-        s3_upload = S3Upload(kind="siae_logo")
-        context["s3_form_values_siae_logo"] = s3_upload.form_values
-        context["s3_upload_config_siae_logo"] = s3_upload.config
-        if self.request.POST:
-            context["label_formset"] = SiaeLabelOldFormSet(self.request.POST, instance=self.object)
-        else:
-            context["label_formset"] = SiaeLabelOldFormSet(instance=self.object)
-        context["last_3_siae_content_filled_full"] = (
-            Siae.objects.with_content_filled_stats()
-            .filter(content_filled_full=True)
-            .exclude(id=self.object.id)
-            .order_by("-updated_at")[:3]
-        )
-        return context
-
-    def post(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        form_class = self.get_form_class()
-        form = self.get_form(form_class)
-        label_formset = SiaeLabelOldFormSet(self.request.POST, instance=self.object)
-        if form.is_valid() and label_formset.is_valid():
-            return self.form_valid(form, label_formset)
-        else:
-            return self.form_invalid(form, label_formset)
-
-    def form_valid(self, form, label_formset):
-        self.object = form.save()
-        label_formset.instance = self.object
-        label_formset.save()
-        return super().form_valid(form)
-
-    def form_invalid(self, form, label_formset):
-        return self.render_to_response(self.get_context_data())
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit_info", args=[self.kwargs.get("slug")])
-
-
-class SiaeEditOfferView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeEditOfferForm
-    template_name = "dashboard/siae_edit_offer.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Vos modifications ont bien été prises en compte."
-
-    def get_context_data(self, **kwargs):
-        """
-        - init forms
-        - pass s3 image upload config
-        """
-        context = super().get_context_data(**kwargs)
-        if self.request.POST:
-            context["offer_formset"] = SiaeOfferFormSet(self.request.POST, instance=self.object)
-            context["client_reference_formset"] = SiaeClientReferenceFormSet(self.request.POST, instance=self.object)
-            context["image_formset"] = SiaeImageFormSet(self.request.POST, instance=self.object)
-        else:
-            context["offer_formset"] = SiaeOfferFormSet(instance=self.object)
-            context["client_reference_formset"] = SiaeClientReferenceFormSet(instance=self.object)
-            context["image_formset"] = SiaeImageFormSet(instance=self.object)
-        s3_upload_client_reference_logo = S3Upload(kind="client_reference_logo")
-        context["s3_form_values_client_reference_logo"] = s3_upload_client_reference_logo.form_values
-        context["s3_upload_config_client_reference_logo"] = s3_upload_client_reference_logo.config
-        s3_upload_siae_image = S3Upload(kind="siae_image")
-        context["s3_form_values_siae_image"] = s3_upload_siae_image.form_values
-        context["s3_upload_config_siae_image"] = s3_upload_siae_image.config
-        return context
-
-    def post(self, request, *args, **kwargs):
-        self.object = self.get_object()
-        form_class = self.get_form_class()
-        form = self.get_form(form_class)
-        offer_formset = SiaeOfferFormSet(self.request.POST, instance=self.object)
-        client_reference_formset = SiaeClientReferenceFormSet(self.request.POST, instance=self.object)
-        image_formset = SiaeImageFormSet(self.request.POST, instance=self.object)
-        if (
-            form.is_valid()
-            and offer_formset.is_valid()
-            and client_reference_formset.is_valid()
-            and image_formset.is_valid()
-        ):
-            return self.form_valid(form, offer_formset, client_reference_formset, image_formset)
-        else:
-            return self.form_invalid(form, offer_formset, client_reference_formset, image_formset)
-
-    def form_valid(self, form, offer_formset, client_reference_formset, image_formset):
-        self.object = form.save()
-        offer_formset.instance = self.object
-        offer_formset.save()
-        client_reference_formset.instance = self.object
-        client_reference_formset.save()
-        image_formset.instance = self.object
-        image_formset.save()
-        return super().form_valid(form)
-
-    def form_invalid(self, form, offer_formset, client_reference_formset, image_formset):
-        return self.render_to_response(self.get_context_data())
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit_offer", args=[self.kwargs.get("slug")])
-
-
-class SiaeEditLinksView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeEditLinksForm
-    template_name = "dashboard/siae_edit_links.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Vos modifications ont bien été prises en compte."
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit_links", args=[self.kwargs.get("slug")])
-
-
-class SiaeEditContactView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeEditContactForm
-    template_name = "dashboard/siae_edit_contact.html"
-    context_object_name = "siae"
-    queryset = Siae.objects.all()
-    success_message = "Vos modifications ont bien été prises en compte."
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_edit_contact", args=[self.kwargs.get("slug")])
-
-
-class SiaeUserRequestConfirmView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeUserRequestForm
-    template_name = "siaes/_siae_user_request_confirm_modal.html"
-    context_object_name = "siaeuserrequest"
-    queryset = SiaeUserRequest.objects.all()
-    success_message = "L'utilisateur a été rattaché à votre structure."
-    # success_url = reverse_lazy("dashboard:siae_users")
-
-    def get_object(self):
-        return get_object_or_404(SiaeUserRequest, id=self.kwargs.get("siaeuserrequest_id"))
-
-    def form_valid(self, form):
-        """
-        - add user to Siae
-        - update SiaeUserRequest
-        - notify user
-        """
-        self.object.siae.users.add(self.object.initiator)
-        self.object.response = True
-        self.object.response_date = timezone.now()
-        self.object.logs.append({"action": "response_true", "timestamp": self.object.response_date.isoformat()})
-        self.object.save()
-        send_siae_user_request_response_email_to_initiator(self.object, response=True)
-        return super().form_valid(form)
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_users", args=[self.kwargs.get("slug")])
-
-
-class SiaeUserRequestCancelView(SiaeMemberRequiredMixin, SuccessMessageMixin, UpdateView):
-    form_class = SiaeUserRequestForm
-    template_name = "siaes/_siae_user_request_cancel_modal.html"
-    context_object_name = "siaeuserrequest"
-    queryset = SiaeUserRequest.objects.all()
-    success_message = "L'utilisateur sera informé de votre refus."
-    # success_url = reverse_lazy("dashboard:siae_users")
-
-    def get_object(self):
-        return get_object_or_404(SiaeUserRequest, id=self.kwargs.get("siaeuserrequest_id"))
-
-    def form_valid(self, form):
-        """
-        - update SiaeUserRequest
-        - notify user
-        """
-        self.object.response = False
-        self.object.response_date = timezone.now()
-        self.object.logs.append({"action": "response_false", "timestamp": self.object.response_date.isoformat()})
-        self.object.save()
-        send_siae_user_request_response_email_to_initiator(self.object, response=False)
-        return super().form_valid(form)
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_users", args=[self.kwargs.get("slug")])
-
-
-class SiaeUserDeleteView(SiaeMemberRequiredMixin, SuccessMessageMixin, DeleteView):
-    template_name = "siaes/_siae_user_delete_modal.html"
-    model = SiaeUser
-    # success_message = "L'utilisateur a été supprimé de votre structure."
-    # success_url = reverse_lazy("dashboard:siae_users")
-
-    def get_object(self):
-        siae = Siae.objects.get(slug=self.kwargs.get("slug"))
-        return get_object_or_404(SiaeUser, siae=siae, id=self.kwargs.get("siaeuser_id"))
-
-    def get_success_url(self):
-        return reverse_lazy("dashboard:siae_users", args=[self.kwargs.get("slug")])
-
-    def get_success_message(self, cleaned_data):
-        return mark_safe(
-            f"L'utilisateur <strong>{self.object.user.full_name}</strong> a été supprimé de votre structure avec succès."  # noqa
-        )
